Book Model and persistence added

+++ src/main/java/wolox/training/repositories/BookRepository.java
@@ -1,5 +1,6 @@
 package wolox.training.repositories;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import wolox.training.models.Book;
 import org.springframework.data.repository.*;
@@ -7,7 +8,7 @@ import org.springframework.data.repository.*;
 import java.util.Optional;
 
 
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 
     Optional<Book> findByAuthor(String author);
